# api_laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Harry Potter",
        "description": "Esta api se encarga de gestionar el contenido de estudiantes, profesores y casas",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/allHouses": {
            "get": {
                "tags": [
                    "Houses"
                ],
                "summary": "Obtener todas las casas de Hogwarts",
                "description": "Devuelve un listado de todas las casas de Hogwarts con sus detalles.",
                "operationId": "429ef1433c7589e8f198be0ca9514736",
                "responses": {
                    "200": {
                        "description": "Un listado de todas las casas de Hogwarts.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "0763e821-09ea-4fc8-bac1-d8fc70c3e4b4"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Hufflepuff"
                                        },
                                        "puntos": {
                                            "type": "integer",
                                            "example": 24
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "https://i.pinimg.com/736x/39/80/92/398092891a39c67b1241734c5fea6843.jpg"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-05-22T07:06:45.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-05-22T07:06:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/house/{id}": {
            "get": {
                "tags": [
                    "Houses"
                ],
                "summary": "Obtiene una casa por su ID",
                "description": "Retorna los detalles de una casa específico de Hogwarts por ID, incluyendo información de los mienbros que pertenecen a este",
                "operationId": "6190e70ffb13abd06565cefbd7e563a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del estudiante",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "0763e821-09ea-4fc8-bac1-d8fc70c3e4b4"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Hufflepuff"
                                                    },
                                                    "puntos": {
                                                        "type": "integer",
                                                        "example": 24
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "https://i.pinimg.com/736x/39/80/92/398092891a39c67b1241734c5fea6843.jpg"
                                                    },
                                                    "student": {
                                                        "type": "string",
                                                        "example": "Hannah Abbott"
                                                    },
                                                    "studentImg": {
                                                        "type": "string",
                                                        "example": null
                                                    },
                                                    "studentId": {
                                                        "type": "string",
                                                        "example": "0a9fe88e-7738-46a7-a8af-592598ce9475"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-05-22T07:06:45.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2024-05-22T07:06:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8000/api/staff?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 9
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8000/api/staff?page=9"
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "url": {
                                                        "type": "string",
                                                        "example": null
                                                    },
                                                    "label": {
                                                        "type": "string",
                                                        "example": "&laquo; Previous"
                                                    },
                                                    "active": {
                                                        "type": "boolean",
                                                        "example": false
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8000/api/staff?page=2"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8000/api/staff"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 12
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": null
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 12
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 23
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Message": {
                                            "type": "string",
                                            "example": "Registro no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}